+++ src/dfx-core/src/error/mod.rs
@@ -5,5 +5,7 @@ pub mod foundation;
 pub mod identity;
 pub mod io;
 pub mod keyring;
+pub mod load_dfx_config;
+pub mod load_networks_config;
 pub mod structured_file;
 pub mod wallet_config;